Create README.md
Using Functions, Classes, Objects, Getters (Accessors) and Setters (Mutators)
Create a Dealership application where you prompt the user to input the make, model, year, and color of the cars they want. Prices will vary based on color, year, model, make. Present the user with different options where they can see the price. Pass all prices for colors and makes to another function which calculates the total prices (with taxes included). If car is black, the user gets a discount of 25% the price of the car; if the car is white, the customer receives a bonus of $400 towards the down payment. If the customer is a war veteran or disabled, they receive 25% off the cost of the car plus $500 bonus. This returns total price and bonuses to the main file. Pass the four values (make, model, year, and color) to another method where you use the values to create objects. Return the objects to the main file where you will display all of them.